requirements.txt
numpy==1.24.0
torch==2.0.1
torchvision==0.15.2
fastapi==0.95.2
uvicorn==0.22.0
pillow==10.0.0
python-multipart==0.0.6
